virt-handler: extract guest-agent command utilities
