+++ pkg/virt-handler/vm.go
@@ -152,38 +152,6 @@ const (
 	memoryHotplugFailedReason = "Memory Hotplug Failed"
 )
 
-var RequiredGuestAgentCommands = []string{
-	"guest-ping",
-	"guest-get-time",
-	"guest-info",
-	"guest-shutdown",
-	"guest-network-get-interfaces",
-	"guest-get-fsinfo",
-	"guest-get-host-name",
-	"guest-get-users",
-	"guest-get-timezone",
-	"guest-get-osinfo",
-}
-
-var SSHRelatedGuestAgentCommands = []string{
-	"guest-ssh-get-authorized-keys",
-	"guest-ssh-add-authorized-keys",
-	"guest-ssh-remove-authorized-keys",
-}
-
-var OldSSHRelatedGuestAgentCommands = []string{
-	"guest-exec-status",
-	"guest-exec",
-	"guest-file-open",
-	"guest-file-close",
-	"guest-file-read",
-	"guest-file-write",
-}
-
-var PasswordRelatedGuestAgentCommands = []string{
-	"guest-set-user-password",
-}
-
 var getCgroupManager = func(vmi *v1.VirtualMachineInstance) (cgroup.Manager, error) {
 	return cgroup.NewManagerFromVM(vmi)
 }
@@ -1518,64 +1486,6 @@ func (c *VirtualMachineController) recordPhaseChangeEvent(vmi *v1.VirtualMachine
 	case v1.Failed:
 		c.recorder.Event(vmi, k8sv1.EventTypeWarning, v1.Stopped.String(), VMICrashed)
 	}
-}
-
-func _guestAgentCommandSubsetSupported(requiredCommands []string, commands []v1.GuestAgentCommandInfo) bool {
-	var found bool
-	for _, cmd := range requiredCommands {
-		found = false
-		for _, foundCmd := range commands {
-			if cmd == foundCmd.Name {
-				if foundCmd.Enabled {
-					found = true
-				}
-				break
-			}
-		}
-		if found == false {
-			return false
-		}
-	}
-	return true
-
-}
-
-func isGuestAgentSupported(vmi *v1.VirtualMachineInstance, commands []v1.GuestAgentCommandInfo) (bool, string) {
-	if !_guestAgentCommandSubsetSupported(RequiredGuestAgentCommands, commands) {
-		return false, "This guest agent doesn't support required basic commands"
-	}
-
-	checkSSH := false
-	checkPasswd := false
-
-	if vmi != nil && vmi.Spec.AccessCredentials != nil {
-		for _, accessCredential := range vmi.Spec.AccessCredentials {
-			if accessCredential.SSHPublicKey != nil && accessCredential.SSHPublicKey.PropagationMethod.QemuGuestAgent != nil {
-				// defer checking the command list so we only do that once
-				checkSSH = true
-			}
-			if accessCredential.UserPassword != nil && accessCredential.UserPassword.PropagationMethod.QemuGuestAgent != nil {
-				// defer checking the command list so we only do that once
-				checkPasswd = true
-			}
-
-		}
-	}
-
-	if checkSSH && !sshRelatedCommandsSupported(commands) {
-		return false, "This guest agent doesn't support required public key commands"
-	}
-
-	if checkPasswd && !_guestAgentCommandSubsetSupported(PasswordRelatedGuestAgentCommands, commands) {
-		return false, "This guest agent doesn't support required password commands"
-	}
-
-	return true, "This guest agent is supported"
-}
-
-func sshRelatedCommandsSupported(commands []v1.GuestAgentCommandInfo) bool {
-	return _guestAgentCommandSubsetSupported(SSHRelatedGuestAgentCommands, commands) ||
-		_guestAgentCommandSubsetSupported(OldSSHRelatedGuestAgentCommands, commands)
 }
 
 func calculatePausedCondition(vmi *v1.VirtualMachineInstance, reason api.StateChangeReason) {
